+++ Songtrain/SoundCloudSongViewController.m
@@ -101,7 +101,7 @@
                 //[self.tracks removeAllObjects];
                 _tracks = (NSArray *)jsonResponse;
                 NSUInteger numReturned = _tracks.count;
-                NSLog(@"Updated favorites to %lu", ((NSArray *)jsonResponse).count);
+                NSLog(@"Updated favorites to %lu", (unsigned long)((NSArray *)jsonResponse).count);
                 dispatch_async(dispatch_get_main_queue(), ^{
                     [weakSelf.wholeTableView reloadData];
                     [weakSelf.wholeTableView.pullToRefreshView stopAnimating];
@@ -160,7 +160,7 @@
             }
         };
         
-        NSString *requestURL = [NSString stringWithFormat:@"%@?offset=%lu&limit=%d", self.location, self.tracks.count, kSoundCloudSongNextLoad];
+        NSString *requestURL = [NSString stringWithFormat:@"%@?offset=%lu&limit=%d", self.location, (unsigned long)self.tracks.count, kSoundCloudSongNextLoad];
         
         [SCRequest performMethod:SCRequestMethodGET onResource:[NSURL URLWithString:requestURL] usingParameters:nil withAccount:[SCSoundCloud account] sendingProgressHandler:nil responseHandler:handler];
     }
